+++ src/test/java/com/cedarsoftware/io/UUIDTest.java
@@ -48,11 +48,11 @@ public class UUIDTest
 
         String json2 = "{\"@type\":\"" + TestUUIDFields.class.getName() + "\", \"internals\": {\"@type\": \"java.util.UUID\", \"leastSigBits\":-7929886640328317609}}";
         thrown = assertThrows(JsonIoException.class, () -> { TestUtil.toObjects(json2, null); });
-        assert thrown.getCause().getMessage().contains("convert from Map to UUID the map must include one of the following: [mostSigBits, leastSigBits], [_v], or [value]");
+        assert thrown.getCause().getMessage().contains("To convert from Map to UUID the map must include one of the following: [UUID], [_v], or [value] with associated values");
 
         String json3 = "{\"@type\":\"" + TestUUIDFields.class.getName() + "\", \"internals\": {\"@type\": \"java.util.UUID\", \"mostSigBits\":7280309849777586861}}";
         thrown = assertThrows(JsonIoException.class, () -> { TestUtil.toObjects(json3, null); });
-        assert thrown.getCause().getMessage().contains("convert from Map to UUID the map must include one of the following: [mostSigBits, leastSigBits], [_v], or [value]");
+        assert thrown.getCause().getMessage().contains("To convert from Map to UUID the map must include one of the following: [UUID], [_v], or [value] with associated values");
     }
 
     @Test